Move `providers` to `data_providers`
