+++ tests/conftest.py
@@ -1,7 +1,6 @@
 
 pytest_plugins = [
     "tests.fixtures.workflow",
-    "tests.fixtures.scope",
     "tests.fixtures.execution",
     "virtool_workflow.testing"
 ]
