+++ virtool_workflow/abc/data_providers/cache.py
