+++ virtool_workflow/analysis/indexes.py
@@ -9,7 +9,7 @@ from virtool_core.utils import decompress_file
 from virtool_workflow import data_model
 
 from virtool_workflow import fixture
-from virtool_workflow.abc.providers.indexes import AbstractIndexProvider
+from virtool_workflow.abc.data_providers.indexes import AbstractIndexProvider
 from virtool_workflow.execution.run_in_executor import FunctionExecutor
 from virtool_workflow.execution.run_subprocess import RunSubprocess
