+++ virtool_workflow/runtime.py
@@ -2,7 +2,7 @@ from typing import Dict, Any
 
 from virtool_workflow.abc import AbstractWorkflowEnvironment
 from virtool_workflow.data_model import Job
-from virtool_workflow.execution.workflow_executor import WorkflowExecution
+from virtool_workflow.execution.workflow_execution import WorkflowExecution
 from virtool_workflow.fixtures.scope import FixtureScope
 from virtool_workflow.workflow import Workflow
 from virtool_workflow_runtime.config.configuration import config_fixtures
